new integration test structure
