+++ integration_test/testUtils.dart
@@ -1,0 +1,9 @@
+import '../lib/main.dart' as app;
+import 'package:shared_preferences/shared_preferences.dart';
+
+Future<void> init() async{
+  SharedPreferences preferences = await SharedPreferences.getInstance();
+  await preferences.clear();
+
+  app.main();
+}
